+++ aiogram_dialog/__init__.py
@@ -1,6 +1,7 @@
 from .api import exceptions
-from .api.events import ChatEvent, Data, ShowMode, StartMode
-from .api.stack import DEFAULT_STACK_ID
+from .api.entities import (
+    ChatEvent, Data, ShowMode, StartMode, DEFAULT_STACK_ID
+)
 from .dialog import Dialog
 from .manager.protocols import BaseDialogManager, DialogManager
 from .manager.registry import DialogRegistry
