+++ aiogram_dialog/api/entities/events.py
@@ -1,0 +1,7 @@
+from typing import Union
+
+from aiogram.types import CallbackQuery, ChatMemberUpdated, Message
+
+from .update_event import DialogUpdateEvent
+
+ChatEvent = Union[CallbackQuery, Message, DialogUpdateEvent, ChatMemberUpdated]
