+++ aiogram_dialog/api/entities/update_event.py
@@ -1,24 +1,27 @@
-from typing import Any, Dict, List, Literal, Optional, Union
+from enum import Enum
+from typing import Any, Optional
 
 from aiogram.fsm.state import State
 from aiogram.types import (
-    CallbackQuery,
     Chat,
-    ChatMemberUpdated,
-    Message,
     TelegramObject,
     Update,
     User,
 )
 
 from .modes import (
-    Action,
     ShowMode,
     StartMode,
 )
 
 DIALOG_EVENT_NAME = "aiogd_update"
-Data = Union[Dict, List, int, str, float, None]
+
+
+class DialogAction(Enum):
+    DONE = "DONE"
+    START = "START"
+    UPDATE = "UPDATE"
+    SWITCH = "SWITCH"
 
 
 class DialogUpdateEvent(TelegramObject):
@@ -31,7 +34,7 @@ class DialogUpdateEvent(TelegramObject):
 
     from_user: User
     chat: Chat
-    action: Action
+    action: DialogAction
     data: Any
     intent_id: Optional[str]
     stack_id: Optional[str]
@@ -47,9 +50,6 @@ class DialogSwitchEvent(DialogUpdateEvent):
     new_state: State
 
 
-ChatEvent = Union[CallbackQuery, Message, DialogUpdateEvent, ChatMemberUpdated]
-
-
 class DialogUpdate(Update):
     aiogd_update: DialogUpdateEvent
 
@@ -63,11 +63,3 @@ class DialogUpdate(Update):
     @property
     def event(self) -> DialogUpdateEvent:
         return self.aiogd_update
-
-
-class FakeUser(User):
-    fake: Literal[True] = True
-
-
-class FakeChat(Chat):
-    fake: Literal[True] = True
