+++ aiogram_dialog/api/internal/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "FakeChat", "FakeUser",
+]
+
+from .fake_data import FakeChat, FakeUser
