+++ aiogram_dialog/api/internal/fake_data.py
@@ -1,0 +1,14 @@
+from typing import Literal
+
+from aiogram.types import (
+    Chat,
+    User,
+)
+
+
+class FakeUser(User):
+    fake: Literal[True] = True
+
+
+class FakeChat(Chat):
+    fake: Literal[True] = True
