+++ aiogram_dialog/api/protocols/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "MediaIdStorageProtocol",
+]
+
+from .media import MediaIdStorageProtocol
