+++ aiogram_dialog/manager/update_handler.py
@@ -1,7 +1,7 @@
 from logging import getLogger
 
-from aiogram_dialog.api.events import (
-    Action,
+from aiogram_dialog.api.entities import (
+    DialogAction,
     DialogStartEvent,
     DialogSwitchEvent,
     DialogUpdateEvent,
@@ -22,7 +22,7 @@ async def handle_update(event: DialogUpdateEvent, dialog_manager: ManagerImpl):
     elif isinstance(event, DialogSwitchEvent):
         await dialog_manager.switch_to(state=event.new_state)
         await dialog_manager.dialog().show()
-    elif event.action is Action.UPDATE:
+    elif event.action is DialogAction.UPDATE:
         if not dialog_manager.current_context():
             logger.warning("No context found")
             return
@@ -30,5 +30,5 @@ async def handle_update(event: DialogUpdateEvent, dialog_manager: ManagerImpl):
             for k, v in event.data.items():
                 dialog_manager.current_context().dialog_data[k] = v
         await dialog_manager.dialog().show()
-    elif event.action is Action.DONE:
+    elif event.action is DialogAction.DONE:
         await dialog_manager.done(result=event.data)
