+++ aiogram_dialog/utils.py
@@ -3,8 +3,8 @@ from typing import Optional, Tuple
 
 from aiogram.types import CallbackQuery, Chat, ChatMemberUpdated, Message, User
 
-from aiogram_dialog.api.events import ChatEvent, DialogUpdateEvent
-from .manager.protocols import MediaId, NewMessage
+from aiogram_dialog.api.entities import ChatEvent, DialogUpdateEvent, MediaId
+from .manager.protocols import NewMessage
 
 logger = getLogger(__name__)
 
@@ -40,11 +40,11 @@ def is_user_loaded(user: User) -> bool:
 
 def get_media_id(message: Message) -> Optional[MediaId]:
     media = (
-        message.audio or
-        message.animation or
-        message.document or
-        (message.photo[-1] if message.photo else None) or
-        message.video
+            message.audio or
+            message.animation or
+            message.document or
+            (message.photo[-1] if message.photo else None) or
+            message.video
     )
     if not media:
         return None
