+++ aiogram_dialog/widgets/kbd/calendar_kbd.py
@@ -5,7 +5,7 @@ from typing import Awaitable, Callable, List, Optional, TypedDict, Union
 
 from aiogram.types import CallbackQuery, InlineKeyboardButton
 
-from aiogram_dialog.api.events import ChatEvent
+from aiogram_dialog.api.entities import ChatEvent
 from aiogram_dialog.manager.protocols import DialogManager, ManagedDialogProto
 from aiogram_dialog.widgets.widget_event import (
     ensure_event_processor,
