+++ aiogram_dialog/widgets/kbd/state.py
@@ -3,7 +3,7 @@ from typing import Any, Callable, Optional
 from aiogram.fsm.state import State
 from aiogram.types import CallbackQuery
 
-from aiogram_dialog.api.events import ChatEvent, Data, StartMode
+from aiogram_dialog.api.entities import ChatEvent, Data, StartMode
 from aiogram_dialog.manager.manager import DialogManager
 from aiogram_dialog.widgets.text import Const, Text
 from aiogram_dialog.widgets.widget_event import WidgetEventProcessor
