+++ aiogram_dialog/widgets/media/static.py
@@ -2,8 +2,8 @@ from typing import Any, Dict, Optional, Union
 
 from aiogram.types import ContentType
 
+from aiogram_dialog.api.entities import MediaAttachment
 from aiogram_dialog.manager.manager import DialogManager
-from aiogram_dialog.manager.protocols import MediaAttachment
 from aiogram_dialog.widgets.text import Const, Text
 from .base import Media
 from ..when import WhenCondition
